[CI] Migrate Eagle 1-GPU tests to test/registered/
Move Eagle 1-GPU tests from test/srt/ to test/registered/eagle/ as part of the CI registry migration.

Changes:
- Move 5 Eagle test files to test/registered/eagle/
- Add CI registry decorators with suite="stage-b-test-small-1-gpu"
- Add stage-b-test-small-1-gpu suite to test/run_suite.py
- Add stage-b-test-small-1-gpu job to pr-test.yml (4 partitions)
- Remove Eagle entries from test/srt/run_suite.py
