+++ test/registered/eagle/test_eagle_infer_beta.py
@@ -1,9 +1,12 @@
 import unittest
+from types import SimpleNamespace
 
 from sglang.srt.environ import envs
 from sglang.srt.utils import kill_process_tree
-from sglang.test.kits.json_constrained_kit import TestJSONConstrainedMixin
-from sglang.test.kits.regex_constrained_kit import TestRegexConstrainedMixin
+from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.few_shot_gsm8k import run_eval
+from sglang.test.kits.matched_stop_kit import MatchedStopMixin
+from sglang.test.kits.radix_cache_server_kit import run_radix_attention_test
 from sglang.test.test_utils import (
     DEFAULT_EAGLE_DRAFT_MODEL_FOR_TEST,
     DEFAULT_EAGLE_TARGET_MODEL_FOR_TEST,
@@ -13,10 +16,10 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
+register_cuda_ci(est_time=194, suite="stage-b-test-small-1-gpu")
 
-class TestEagleConstrainedDecoding(
-    CustomTestCase, TestRegexConstrainedMixin, TestJSONConstrainedMixin
-):
+
+class TestEagleServerBase(CustomTestCase, MatchedStopMixin):
     max_running_requests = 64
     attention_backend = "triton"
     spec_steps = 5
@@ -26,8 +29,6 @@ class TestEagleConstrainedDecoding(
     other_launch_args = []
     model = DEFAULT_EAGLE_TARGET_MODEL_FOR_TEST
     draft_model = DEFAULT_EAGLE_DRAFT_MODEL_FOR_TEST
-    grammar_backend = "xgrammar"
-    eagle_v2 = False
 
     @classmethod
     def setUpClass(cls):
@@ -52,11 +53,11 @@ class TestEagleConstrainedDecoding(
             "0.75",
             "--max-running-requests",
             str(cls.max_running_requests),
-            "--grammar-backend",
-            cls.grammar_backend,
+            "--cuda-graph-bs",
+            *[str(i) for i in range(1, cls.max_running_requests + 1)],
         ]
         launch_args.extend(cls.other_launch_args)
-        with envs.SGLANG_ENABLE_SPEC_V2.override(cls.eagle_v2):
+        with envs.SGLANG_ENABLE_SPEC_V2.override(True):
             cls.process = popen_launch_server(
                 cls.model,
                 cls.base_url,
@@ -68,9 +69,30 @@ class TestEagleConstrainedDecoding(
     def tearDownClass(cls):
         kill_process_tree(cls.process.pid)
 
+    def test_radix_attention(self):
+        run_radix_attention_test(self.base_url)
+        assert self.process.poll() is None
 
-class TestEagleConstrainedDecodingV2(TestEagleConstrainedDecoding):
-    eagle_v2 = True
+    def test_gsm8k(self):
+        args = SimpleNamespace(
+            num_shots=5,
+            data_path=None,
+            num_questions=1000,
+            max_new_tokens=512,
+            parallel=128,
+            host="http://127.0.0.1",
+            port=int(self.base_url.split(":")[-1]),
+        )
+        metrics = run_eval(args)
+        print(f"TestEagleLargeBS -- {metrics=}")
+        self.assertGreater(
+            metrics["accuracy"], 0.23
+        )  # 0.3333 for 60 questions; 0.234 for 1319 questions
+        assert self.process.poll() is None
+
+
+class TestEagleServerPage(TestEagleServerBase):
+    other_launch_args = ["--page-size", "64"]
 
 
 if __name__ == "__main__":
