+++ test/run_suite.py
@@ -17,7 +17,7 @@ HW_MAPPING = {
 PER_COMMIT_SUITES = {
     HWBackend.CPU: ["default"],
     HWBackend.AMD: ["stage-a-test-1"],
-    HWBackend.CUDA: ["stage-a-test-1"],
+    HWBackend.CUDA: ["stage-a-test-1", "stage-b-test-small-1-gpu"],
     HWBackend.NPU: [],
 }
